Refactoring way how upgrades are handled on recurring payment
This refactoring was necessary to properly handle payment items generated
on charges after the upgrade.

Original implementation used custom_amount and updated
"subscription_type_id"for all upgrade charges which was not correct.
The use of next_subscription_type_id should be default scenario,
and custom_amount was meant to be used only for exceptions.

Once the upgrade is handled via next subscription, the recurrent payment
then generates new items and calculates the price based on the
next subscription type.
If the upgrade is handled via custom_amount and more than one item is
defined on the subscription type, system cannot generate items for
payments and such charge ends up with error.

In upgrader, two major changes happened:

- futureChargeAmount was renamed to customAmount to indicate that only
"exceptions" in charges should be set here
- futureChargeAmount() now returns either customAmount or standard
price of subscription type we're upgrading to

remp/crm#472